pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "kwpbridge"
version = "0.1.0"
description = "KWP1281 / K-Line diagnostic bridge for VAG vehicles"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov"]

[project.scripts]
kwpbridge = "kwpbridge.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["kwpbridge*"]
